+++ tests/EwayClientTest.php
@@ -3,16 +3,18 @@
 namespace Bdt\Test\Eway;
 
 use Bdt\Eway\EwayClient;
-use Guzzle\Tests\GuzzleTestCase;
-use Guzzle\Plugin\Log\LogPlugin;
+use GuzzleHttp\Command\Event\CommandEvents;
+use GuzzleHttp\Subscriber\Mock;
 
-class EwayClientTest extends GuzzleTestCase
+class EwayClientTest extends \PHPUnit_Framework_TestCase
 {
     public function setup()
     {
         $this->client = EwayClient::factory(array(
             'base_url' => 'https://www.eway.com.au/gateway_cvn/xmltest/testpage.asp',
         ));
+        $this->mock = new Mock;
+        $this->client->getHttpClient()->getEmitter()->attach($this->mock);
     }
 
     public function testCanSetCustomerIdInClient()
@@ -29,14 +31,14 @@ class EwayClientTest extends GuzzleTestCase
             'cardExpiryYear'  => '20',
             'CVN'             => '123',
         ));
-        $command->prepare();
+        CommandEvents::prepare($command, $this->client);
         $this->assertEquals(100, $command['customerID']);
     }
 
     public function testSendPayment()
     {
-        $this->setMockResponse($this->client, 'sendpayment_success.txt');
-        $response = $this->client->getCommand('SendPayment', array(
+        $this->setMockResponse('sendpayment_success.txt');
+        $response = $this->client->SendPayment(array(
             'customerID'      => '87654321',
             'totalAmount'     => '10',
             'cardHoldersName' => 'Foo Bar',
@@ -44,16 +46,16 @@ class EwayClientTest extends GuzzleTestCase
             'cardExpiryMonth' => '06',
             'cardExpiryYear'  => '20',
             'CVN'             => '123',
-        ))->execute();
+        ));
         $this->assertEquals('10', $response['trxnError']['code']);
         $this->assertEquals('Approved For Partial Amount(Test CVN Gateway)', $response['trxnError']['message']);
-        $this->assertTrue($response->get('trxnStatus'));
+        $this->assertTrue($response['trxnStatus']);
     }
 
     public function testSendPaymentFail()
     {
-        $this->setMockResponse($this->client, 'sendpayment_failure.txt');
-        $response = $this->client->getCommand('SendPayment', array(
+        $this->setMockResponse('sendpayment_failure.txt');
+        $response = $this->client->SendPayment(array(
             'customerID'      => '87654321',
             'totalAmount'     => '13',
             'cardHoldersName' => 'Foo Bar',
@@ -61,7 +63,7 @@ class EwayClientTest extends GuzzleTestCase
             'cardExpiryMonth' => '06',
             'cardExpiryYear'  => '20',
             'CVN'             => '123',
-        ))->execute();
+        ));
         $this->assertEquals('13', $response['trxnError']['code']);
         $this->assertEquals('Invalid Amount(Test CVN Gateway)', $response['trxnError']['message']);
         $this->assertFalse($response['trxnStatus']);
@@ -69,21 +71,27 @@ class EwayClientTest extends GuzzleTestCase
 
     public function testSendPaymentFail2()
     {
-        $this->setMockResponse($this->client, 'sendpayment_failure_2.txt');
-        $response = $this->client->getCommand('SendPayment', array(
+        $this->setMockResponse('sendpayment_failure_2.txt');
+        $response = $this->client->SendPayment(array(
             'customerID'      => '87654321',
             'totalAmount'     => '13',
             'cardHoldersName' => 'Foo Bar',
-            'cardNumber'      => '4111111111111111',
+            'cardNumber'      => '4444333322221111',
             'cardExpiryMonth' => '06',
             'cardExpiryYear'  => '20',
             'CVN'             => '123',
-        ))->execute();
+        ));
         $this->assertEquals(null, $response['trxnError']['code']);
         $this->assertEquals('eWAY Error: Invalid Expiry Date. Your credit card has not been  billed for this transaction.',
             $response['trxnError']['message']
         );
         $this->assertFalse($response['trxnStatus']);
+    }
+
+    private function setMockResponse($file)
+    {
+        $file = file_get_contents(__DIR__.'/mock/'.$file);
+        $this->mock->addResponse($file);
     }
 
 }